+++ PreviewCode/AppDelegateFonts.swift
@@ -73,15 +73,7 @@ extension AppDelegate {
         // All done, update the main stores and begin to load
         // settings (which immediately updates the UI, via `displaySettings()`,
         // which itself requires the font store to be populated
-        /*
-        DispatchQueue.main.async {
-            self.fonts = cf
-            self.loadSettings()
-        }
-        */
-
-        // FROM 2.2.4
-        // Switch to Swift Concurrency
+        // FROM 2.2.4 - Switch to Swift Concurrency
         Task {
             // Run task on main thread (when it's free?)
             await MainActor.run {