+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 
 - 2.2.4 *Unreleased*
     - Update to `HighlighterSwift` 3.0.1.
+    - Migrate to Swift Concurrency.
     - Correctly disable tab switching when sending feedback.
+    - Fix truncation of feedback text to prevent loss of styling.
 - 2.2.3 *12 December 2025*
     - Add `.LIS` as a Basic file extension for we Amstrad NC100 users.
     - Add `.TPL` as an OPL file extension for Psion users.